# Remove IAssemblyName (and various fusion remnants)