--- a/src/coreclr/binder/assemblybinder.cpp
+++ b/src/coreclr/binder/assemblybinder.cpp
@@ -20,18 +20,12 @@
 #include "bindresult.inl"
 #include "failurecache.hpp"
 #include "utils.hpp"
-#include "variables.hpp"
 #include "stringarraylist.h"
 #include "configuration.h"
-
-#define APP_DOMAIN_LOCKED_UNLOCKED        0x02
-#define APP_DOMAIN_LOCKED_CONTEXT         0x04
 
 #ifndef IMAGE_FILE_MACHINE_ARM64
 #define IMAGE_FILE_MACHINE_ARM64             0xAA64  // ARM64 Little-Endian
 #endif
-
-BOOL IsCompilationProcess();
 
 #if !defined(DACCESS_COMPILE) && !defined(CROSSGEN_COMPILE)
 #include "clrprivbindercoreclr.h"
@@ -113,12 +107,13 @@ namespace BINDER_SPACE
             return true;
         }
 
+        const WCHAR* s_httpURLPrefix = W("http://");
         HRESULT URLToFullPath(PathString &assemblyPath)
         {
             HRESULT hr = S_OK;
 
             SString::Iterator pos = assemblyPath.Begin();
-            if (assemblyPath.MatchCaseInsensitive(pos, g_BinderVariables->httpURLPrefix))
+            if (assemblyPath.MatchCaseInsensitive(pos, s_httpURLPrefix))
             {
                 // HTTP downloads are unsupported
                 hr = FUSION_E_CODE_DOWNLOAD_DISABLED;
@@ -186,23 +181,6 @@ namespace BINDER_SPACE
         }
 #endif // !CROSSGEN_COMPILE
     };
-
-    /* static */
-    HRESULT AssemblyBinder::Startup()
-    {
-        STATIC_CONTRACT_NOTHROW;
-
-        HRESULT hr = S_OK;
-
-        // This should only be called once
-        _ASSERTE(g_BinderVariables == NULL);
-        g_BinderVariables = new Variables();
-        IF_FAIL_GO(g_BinderVariables->Init());
-
-    Exit:
-        return hr;
-    }
-
 
     HRESULT AssemblyBinder::TranslatePEToArchitectureType(DWORD  *pdwPAFlags, PEKIND *PeKind)
     {
@@ -373,9 +351,6 @@ namespace BINDER_SPACE
                                          Assembly **ppSystemAssembly,
                                          bool       fBindToNativeImage)
     {
-        // Indirect check that binder was initialized.
-        _ASSERTE(g_BinderVariables != NULL);
-
         HRESULT hr = S_OK;
 
         _ASSERTE(ppSystemAssembly != NULL);
@@ -469,9 +444,6 @@ namespace BINDER_SPACE
         SString& cultureName,
         Assembly** ppSystemAssembly)
     {
-        // Indirect check that binder was initialized.
-        _ASSERTE(g_BinderVariables != NULL);
-
         HRESULT hr = S_OK;
 
         _ASSERTE(ppSystemAssembly != NULL);
@@ -782,8 +754,7 @@ namespace BINDER_SPACE
 
         if (!tpaListAssembly)
         {
-            SString &culture = pRequestedAssemblyName->GetCulture();
-            if (culture.IsEmpty() || culture.EqualsCaseInsensitive(g_BinderVariables->cultureNeutral))
+            if (pRequestedAssemblyName->IsNeutralCulture())
             {
                 dwIncludeFlags |= AssemblyName::EXCLUDE_CULTURE;
             }
@@ -916,7 +887,7 @@ namespace BINDER_SPACE
             SString& simpleNameRef = pRequestedAssemblyName->GetSimpleName();
             SString& cultureRef = pRequestedAssemblyName->GetCulture();
 
-            _ASSERTE(!cultureRef.IsEmpty() && !cultureRef.EqualsCaseInsensitive(g_BinderVariables->cultureNeutral));
+            _ASSERTE(!pRequestedAssemblyName->IsNeutralCulture());
 
             ReleaseHolder<Assembly> pAssembly;
             SString fileName;
@@ -1046,10 +1017,9 @@ namespace BINDER_SPACE
     {
         HRESULT hr = S_OK;
 
-        SString &culture = pRequestedAssemblyName->GetCulture();
         bool fPartialMatchOnTpa = false;
 
-        if (!culture.IsEmpty() && !culture.EqualsCaseInsensitive(g_BinderVariables->cultureNeutral))
+        if (!pRequestedAssemblyName->IsNeutralCulture())
         {
             IF_FAIL_GO(BindSatelliteResource(pApplicationContext, pRequestedAssemblyName, pBindResult));
         }
@@ -1466,9 +1436,6 @@ HRESULT AssemblyBinder::BindUsingPEImage(/* in */  ApplicationContext *pApplicat
                                          /* [retval] [out] */  Assembly **ppAssembly)
 {
     HRESULT hr = E_FAIL;
-
-    // Indirect check that binder was initialized.
-    _ASSERTE(g_BinderVariables != NULL);
 
     LONG kContextVersion = 0;
     BindResult bindResult;