--- a/src/coreclr/binder/assemblyname.cpp
+++ b/src/coreclr/binder/assemblyname.cpp
@@ -22,6 +22,7 @@
 
 namespace
 {
+    // See https://docs.microsoft.com/dotnet/framework/reflection-and-codedom/specifying-fully-qualified-type-names#specifying-assembly-names
     const WCHAR* s_neutralCulture = W("neutral");
 }