--- a/src/coreclr/binder/clrprivbindercoreclr.cpp
+++ b/src/coreclr/binder/clrprivbindercoreclr.cpp
@@ -4,7 +4,6 @@
 #include "common.h"
 #include "assemblybinder.hpp"
 #include "clrprivbindercoreclr.h"
-#include "variables.hpp"
 
 using namespace BINDER_SPACE;