--- a/src/coreclr/binder/inc/assembly.hpp
+++ b/src/coreclr/binder/inc/assembly.hpp
@@ -72,7 +72,7 @@ namespace BINDER_SPACE
         LPCWSTR GetSimpleName();
 
         STDMETHOD(BindAssemblyByName)(
-            /* [in] */ struct AssemblyNameData *pAssemblyNameData,
+            /* [in] */ AssemblyNameData *pAssemblyNameData,
             /* [retval][out] */ ICLRPrivAssembly **ppAssembly);
 
         STDMETHOD(GetAvailableImageTypes)(PDWORD pdwImageTypes);