Remove a bunch of remnants of fusion from the binder. Definitely lots more that this change doesn't deal with (e.g. the existence of the `ICLRPrivBinder` interface - it is internal only and our usage of if it already doesn't adhere to COM semantics).
- Replace `IAssemblyName` with a simple struct (`AssemblyNameData`) for passing information to binder instances. (If we get rid of the interface per above, we can remove the intermediary struct as well - but it is at least simpler and less costly than `IAssemblyName`.)
- Delete a bunch of files / functions that were just for dealing with `IAssemblyName`
- Move some types from deleted files to bindertypes.hpp: `PEKIND` from peinformation.h, `AssemblyContentType` and `ASM_DISPLAY_FLAGS` from fusion.h

This avoids re-parsing of the display name and some data copying, so it actually improves perf some too. Relative to the cost of actually loading an assembly, it is small / lost in the noise, so I tried narrowing down a measurement to the cost of just going through bind resolution by using custom ALCs and calling `LoadFromAssemblyName` for an already loaded assembly.

Stopwatch timings for 150k iterations using my local build:
|| Before - 150k iterations (s) | After - 150k iterations (s) |
|---|---:|---:|
| Returned by ALC.Load | 1.8 | 0.9 |
| Fall back to default ALC | 1.9 | 1.0 |

<details>
<summary>Test code</summary>

```C#
Assembly asm = Assembly.GetExecutingAssembly();
AssemblyName asmName = asm.GetName();
int iterations = 150000;
AssemblyLoadContext[] alcs = new AssemblyLoadContext[iterations];
for (int i = 0; i < alcs.Length; i++)
{
    alcs[i] = ...
}

Stopwatch stopwatch = new Stopwatch();
stopwatch.Start();
for (int i = 0; i < alcs.Length; i++)
{
    alcs[i].LoadFromAssemblyName(asmName);
}
stopwatch.Stop();
```

For returning assembly in Load:
```C#
alcs[i] = new CustomALC()
...
class CustomALC : AssemblyLoadContext
{
    protected override Assembly Load(AssemblyName assemblyName)
    {
        return assemblyName.Name == asmName.Name ? asm : null;
    }
}
```

For fallling back to default ALC:
```C#
alcs[i] = new AssemblyLoadContext("DefaultALCFallback")
```
</details>